filenet/requirements.txt
flask==2.2.3
werkzeug==2.2.3
flask-login==0.6.2
flask-wtf==1.1.1
pymongo==4.3.3
boto3==1.26.135
python-dotenv==1.0.0
bcrypt==4.0.1
flask-session==0.4.0
gunicorn==20.1.0
email-validator==2.0.0
pytest==7.3.1
flask-restful==0.3.9
